refactor:create one interface CrudOperations
