+++ src/main/java/ru/practicum/shareit/exception/userException/UserAlreadyExistException.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.exception;
+package ru.practicum.shareit.exception.userException;
 
 public class UserAlreadyExistException extends RuntimeException {
     public UserAlreadyExistException(String message) {
