+++ src/main/java/ru/practicum/shareit/exception/userException/UserNotExistException.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.exception;
+package ru.practicum.shareit.exception.userException;
 
 public class UserNotExistException extends RuntimeException {
     public UserNotExistException(String message) {
