+++ src/main/java/ru/practicum/shareit/exception/userException/UserValidationException.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.exception;
+package ru.practicum.shareit.exception.userException;
 
 public class UserValidationException extends RuntimeException {
     public UserValidationException(String message) {
