--- src/main/java/ru/practicum/shareit/storage/UserStorage.java
@@ -1,6 +1,0 @@
-package ru.practicum.shareit.storage;
-
-import ru.practicum.shareit.user.model.User;
-
-public interface UserStorage extends Storage<User> {
-}
