+++ src/main/java/ru/practicum/shareit/user/storage/UserStorage.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.user.storage;
+
+import ru.practicum.shareit.CrudOperations;
+import ru.practicum.shareit.user.model.User;
+
+public abstract class UserStorage implements CrudOperations<User> {
+}
