+++ src/test/java/ru/practicum/shareit/user/UserStorageTest.java
@@ -1,7 +1,7 @@
 package ru.practicum.shareit.user;
 
 import org.junit.jupiter.api.Test;
-import ru.practicum.shareit.storage.UserStorage;
+import ru.practicum.shareit.user.storage.UserStorage;
 import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.user.storage.InMemoryUserStorage;
 import static org.junit.jupiter.api.Assertions.assertEquals;
